Made Changes to How Firebase App is initialized.
I created a public constructor, which is called on server startup, that initializes the Firebase App. Made the Firebase App initialize with an app name, openfireAuth, so that it can co-exist with the Firebase Apps in my Ktor Plugin. If you do not provide a name, it defaults to "DEFAULT", and two Firebase Apps cannot have the same name. Added a check in the initFirebase() method to make sure that an instance of the "openfireAuth" does not already exist. If it doesn't exist, one will be created. I moved the JSON credentials needed to create a FirebaseApp into the "resource" folder of the project so that I can get a reference to the file without having the full file path. This ensures that it will work on Linux as well.